cleanCache
